chore: split out tests from parser module
